omap3evm: ethernet: split function setup_net_chip
In current implementation, the function sets up the
ethernet chip and resets it - based on EVM revision.

Board revision can be detected only by probing the
ethernet chip. However, steps to reset the chip can
be excluded if support for ethernet is not required.

Steps to reset the ethernet chip are moved to a new
function - reset_net_chip().

Signed-off-by: Sanjeev Premi <premi@ti.com>