omap3evm: quickboot: Exclude IVA initialization
It is not essential to initialize the IVA subsystem
in the u-boot. But, it is possible that certain OS
versions depending upon u-boot for the same.

Code and data related to the initialization of IVA
is being wrapped in #ifdef CONFIG_OMAP_IVA...#endif.

To maintain backward compatibility, CONFIG_OMAP_IVA
is defined in the default confoguration for the
OMAP3EVM as well.