drivers: net: cpsw: reset slave ports on device close
When the cpsw device is closed, we need to cleanly shutdown
the dma state and the slave ports. With out this, we were
seeing failures(on subsequent open and later when kernel boots up)
if the device close function is invoked under heavy traffic

Fixed this by adding slave port reset on device close.

Signed-off-by: Sriramakrishnan A G <srk@ti.com>